+++ build.gradle
@@ -2,7 +2,8 @@ plugins {
     id 'org.springframework.boot' version '2.1.0.RELEASE'
     id "org.jetbrains.kotlin.kapt" version "1.3.50"
     id 'org.jetbrains.kotlin.jvm' version '1.3.50'
-    id 'org.jetbrains.kotlin.plugin.allopen' version '1.3.50'
+    id 'org.jetbrains.kotlin.plugin.spring' version '1.3.50'
+    id 'org.jetbrains.kotlin.plugin.jpa' version '1.3.50'
     id 'org.jetbrains.kotlin.plugin.noarg' version '1.3.50'
     id 'org.hidetake.swagger.generator' version '2.18.1'
     id "io.spring.dependency-management" version "1.0.9.RELEASE"
@@ -68,4 +69,8 @@ compileKotlin {
 }
 compileTestKotlin {
     kotlinOptions.jvmTarget = "1.8"
+}
+
+sourceSets {
+    main.java.srcDirs += 'src/main/kotlin'
 }
