+++ src/main/kotlin/ru/tinkoff/fintech/service/transaction/TransactionService.kt
@@ -1,0 +1,8 @@
+package ru.tinkoff.fintech.service.transaction
+
+import ru.tinkoff.fintech.model.Transaction
+
+interface TransactionService {
+
+    fun handle(transaction: Transaction)
+}
